Use validator subclasses for question and answers
