+++ app/validation/questions/question_validator.py
@@ -1,0 +1,9 @@
+from app.validation.validator import Validator
+
+
+class QuestionValidator(Validator):
+    question = {}
+
+    def __init__(self, schema_element):
+        super().__init__(schema_element)
+        self.question = schema_element
