+++ tests/test_date_range_question_validator.py
@@ -1,56 +1,6 @@
-from app.validation.question_validator import QuestionValidator
-
-
-def test_invalid_id_in_answers_to_calculate():
-    question = {
-        "answers": [
-            {
-                "decimal_places": 2,
-                "id": "breakdown-1",
-                "label": "Breakdown 1",
-                "type": "Number",
-            },
-            {
-                "decimal_places": 2,
-                "id": "breakdown-2",
-                "label": "Breakdown 2",
-                "type": "Number",
-            },
-        ],
-        "calculations": [
-            {
-                "answer_id": "total-answer",
-                "answers_to_calculate": [
-                    "breakdown-1",
-                    "breakdown-2",
-                    "breakdown-3",
-                    "breakdown-4",
-                ],
-                "calculation_type": "sum",
-                "conditions": ["equals"],
-            }
-        ],
-        "id": "breakdown-question",
-        "title": "Breakdown",
-        "type": "Calculated",
-    }
-    question_validator = QuestionValidator(question)
-    question_validator.validate_answers_to_calculate()
-
-    expected_error_messages = [
-        {
-            "message": "Answer does not exist within this question",
-            "id": "breakdown-question",
-            "answer_id": "breakdown-3",
-        },
-        {
-            "message": "Answer does not exist within this question",
-            "id": "breakdown-question",
-            "answer_id": "breakdown-4",
-        },
-    ]
-
-    assert expected_error_messages == question_validator.errors
+from app.validation.questions.date_range_question_validator import (
+    DateRangeQuestionValidator,
+)
 
 
 def test_invalid_date_range():
@@ -78,8 +28,8 @@ def test_invalid_date_range():
         "type": "DateRange",
     }
 
-    question_validator = QuestionValidator(question)
-    question_validator.validate_date_range()
+    question_validator = DateRangeQuestionValidator(question)
+    question_validator.validate()
 
     expected_error_messages = [
         {
@@ -121,8 +71,8 @@ def test_invalid_yyyy_date_range_period():
         "type": "DateRange",
     }
 
-    question_validator = QuestionValidator(question)
-    question_validator.validate_date_range()
+    question_validator = DateRangeQuestionValidator(question)
+    question_validator.validate()
 
     expected_error_messages = [
         {
@@ -167,8 +117,8 @@ def test_invalid_mm_yyyy_date_range_period():
         "type": "DateRange",
     }
 
-    question_validator = QuestionValidator(question)
-    question_validator.validate_date_range()
+    question_validator = DateRangeQuestionValidator(question)
+    question_validator.validate()
 
     expected_error_messages = [
         {
